Move tests and rename files to be more logical
