+++ src/app/common/structs/mod.rs
@@ -16,7 +16,7 @@ mod locality;
 mod name;
 mod postal_code;
 mod previous_election_results;
-mod problems;
+mod problematic;
 mod utc_date_time;
 
 pub use address::{Address, DutchAddress, InternationalAddress};
@@ -37,5 +37,5 @@ pub use locality::{Locality, PlaceOfResidence};
 pub use name::FullName;
 pub use postal_code::{InternationalPostalCode, PostalCode};
 pub use previous_election_results::PreviousElectionResults;
-pub use problems::{PotentialProblems, Problematic, Severity};
+pub use problematic::{PotentialProblems, Problematic, Severity};
 pub use utc_date_time::UtcDateTime;
